feat: add prometheus relation
This adds a `prometheus` relation that consumes the grafana-source interface so that we can connect this charm to the Prometheus charm

This integration is implemented using the GrafanaSourceConsumer library, which is probably overpowered for what we need here.  A future implementation might be better to use a simpler interface or a custom implementation that does not require a peer relation and other details.