feat: use prometheus-api to relate to prometheus
Previously, we used the grafana-source lib/interface to relate to Prometheus.  This had a few drawbacks, mainly that GrafanaSource needed a peer relation and did a lot of grafana-specific stuff that we didn't need.

This commit removes the grafana-source integration, replacing it with the prometheus-api integration.  Kiali still works the same as always, the only change here is how it receives the url for prometheus.

Note that this is a breaking change in that previously kiali's `prometheus` relation supported grafana_source as the interface, and now it supports prometheus_api.