feat: use a context manager to compute status from charm reconciliation
Previously, every step in `reconcile` captured the errors emitted and generated charm statuses for the different parts.  The charm status was then computed as the worst of these statuses.

This commit refactors that to use a context manager (StatusManager) to offload this work of catching known errors and mapping them to charm statuses. This reduces the boilerplate code in `reconcile`.