refactor: remove collect_status event handler, handle statuses in reconcile
Previously, unit status was handled in a collect_status event handler.  This meant that as we increased complexity of the reconcile function a lot of logic would need to be duplicated in the collect_status handler in order to give users a nuanced view of the charm status.

To address this issue, the current commit refactors status handling to occur completely in the reconcile function.  To achieve this, the helpers called in `reconcile` will all raise custom exceptions on known errors.  These exceptions can be mapped to the desired charm statuses.  The benefit of this pattern is that we can embed the logic of "errorX should generate statusMessageY" in helpers without having helpers directly setting charm status