refactor: clean up how configure_kiali_workload handles null input (#56)
Previously, configure_kiali_workload raised Blocked on the case where new_config=None, but this isn't really an error so much as a request that we shut down kiali.  This has been refactored so we do not Block for that reason.

Also refactored here is how we handle when the service has not yet been created.  Instead of suppressing an APIError, we instead check for a model and suppress a ModelError as its less likely to accidentally suppress some unrelated error.